Replace iconv with transliterate
iconv sometimes replaces non-ASCII characters with a question mark, for example
the character Ë. I cannot reproduce this in a test: the unit tests pass perfectly.
However, the problem occurs when calling the function in a web environment.
The function `transliterator_transliterate` does not seem to have the problem.
I've added a few extra non-ASCII characters to the unit tests to make sure.